- Move namespaces removal method from utils to netconf plugin
