+++ lib/ansible/plugins/netconf/iosxr.py
@@ -22,14 +22,16 @@ __metaclass__ = type
 
 import json
 import re
+import sys
 import collections
+from io import BytesIO
+from ansible.module_utils.six import StringIO
 
 from ansible import constants as C
-from ansible.module_utils.iosxr import build_xml, remove_namespaces
+from ansible.module_utils.iosxr import build_xml
 from ansible.errors import AnsibleConnectionFailure, AnsibleError
 from ansible.plugins.netconf import NetconfBase
 from ansible.plugins.netconf import ensure_connected
-from lxml import etree
 
 try:
     from ncclient import manager
@@ -39,12 +41,50 @@ try:
 except ImportError:
     raise AnsibleError("ncclient is not installed")
 
+try:
+    from lxml import etree
+except ImportError:
+    raise AnsibleError("lxml is not installed")
 
-def remove_ns(response):
-    ret = remove_namespaces(response)
-    if ret == 'no_lxml':
-        raise AnsibleError('lxml not installed')
-    return ret
+
+def transform_reply():
+    reply = '''<xsl:stylesheet version="1.0" xmlns:xsl="http://www.w3.org/1999/XSL/Transform">
+    <xsl:output method="xml" indent="no"/>
+
+    <xsl:template match="/|comment()|processing-instruction()">
+        <xsl:copy>
+            <xsl:apply-templates/>
+        </xsl:copy>
+    </xsl:template>
+
+    <xsl:template match="*">
+        <xsl:element name="{local-name()}">
+            <xsl:apply-templates select="@*|node()"/>
+        </xsl:element>
+    </xsl:template>
+
+    <xsl:template match="@*">
+        <xsl:attribute name="{local-name()}">
+            <xsl:value-of select="."/>
+        </xsl:attribute>
+    </xsl:template>
+    </xsl:stylesheet>
+    '''
+    if sys.version < '3':
+        return reply
+    else:
+        print("utf8")
+        return reply.encode('UTF-8')
+
+
+# Note: Workaround for ncclient 0.5.3
+def remove_namespaces(rpc_reply):
+    xslt = transform_reply()
+    parser = etree.XMLParser(remove_blank_text=True)
+    xslt_doc = etree.parse(BytesIO(xslt), parser)
+    transform = etree.XSLT(xslt_doc)
+
+    return etree.fromstring(str(transform(etree.parse(StringIO(str(rpc_reply))))))
 
 
 class Netconf(NetconfBase):
@@ -128,7 +168,7 @@ class Netconf(NetconfBase):
     def get(self, *args, **kwargs):
         try:
             response = self.m.get(*args, **kwargs)
-            return to_xml(remove_ns(response))
+            return to_xml(remove_namespaces(response))
         except RPCError as exc:
             raise Exception(to_xml(exc.xml))
 
@@ -136,7 +176,7 @@ class Netconf(NetconfBase):
     def get_config(self, *args, **kwargs):
         try:
             response = self.m.get_config(*args, **kwargs)
-            return to_xml(remove_ns(response))
+            return to_xml(remove_namespaces(response))
         except RPCError as exc:
             raise Exception(to_xml(exc.xml))
 
@@ -144,7 +184,7 @@ class Netconf(NetconfBase):
     def edit_config(self, *args, **kwargs):
         try:
             response = self.m.edit_config(*args, **kwargs)
-            return to_xml(remove_ns(response))
+            return to_xml(remove_namespaces(response))
         except RPCError as exc:
             raise Exception(to_xml(exc.xml))
 
@@ -152,7 +192,7 @@ class Netconf(NetconfBase):
     def commit(self, *args, **kwargs):
         try:
             response = self.m.commit(*args, **kwargs)
-            return to_xml(remove_ns(response))
+            return to_xml(remove_namespaces(response))
         except RPCError as exc:
             raise Exception(to_xml(exc.xml))
 
@@ -160,7 +200,7 @@ class Netconf(NetconfBase):
     def validate(self, *args, **kwargs):
         try:
             response = self.m.validate(*args, **kwargs)
-            return to_xml(remove_ns(response))
+            return to_xml(remove_namespaces(response))
         except RPCError as exc:
             raise Exception(to_xml(exc.xml))
 
@@ -168,6 +208,6 @@ class Netconf(NetconfBase):
     def discard_changes(self, *args, **kwargs):
         try:
             response = self.m.discard_changes(*args, **kwargs)
-            return to_xml(remove_ns(response))
+            return to_xml(remove_namespaces(response))
         except RPCError as exc:
             raise Exception(to_xml(exc.xml))
